Add support to apply patches before calculating size
This should ideally be not needed as we should be able to control all
the configurations via config files. OTA currently defines the logging
level in a header file as opposed to taking it from a cofig file. As a
result, we need to patch the header file before calculating the size as
we want to turn off logging while calculating size. Later when logging
level is moved to a config file, it will not be needed and we will turn
off logging the header file contained with the memory estimator tool.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>