Add reconnection and resubscribe logic to the MQTT lightweight, keep alive, and mutual auth demos (#291)
This PR simply follows changes from PR #271: It copies changes from the plaintext demo and implements retries for the MQTT lightweight, keep alive, and mutual auth demos.

If a connect or subscribe attempt fails, vTaskDelay is called to make the task sleep for some bounded backoff period that doubles on each retry. In addition, SUBACK status codes are updated to support multiple topic filters. This way, if a SUBSCRIBE is sent for multiple topic filters and the broker returns 0x80 (a failure) for any of them, a SUBSCRIBE will attempt to be resent.