Fix send zero bytes stream buffer unit test
The change in https://github.com/FreeRTOS/FreeRTOS-Kernel/pull/264
updated prvWriteMessageToBuffer in a way that it no longer calls
prvWriteBytesToBuffer for messages of zero size (prvWriteBytesToBuffer
triggers an assert for messages of zero size). As a result, a zero byte
send to a stream buffer no longer results in assert failure but instead
returns 0. This change updates the zero byte send test to check for 0
return value.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>