Update Tasks.c CBMC Proofs to Latest Code
* Fix Tasks.c patch, line numbers were out of sync and patching was
broken.
* Add assumption to TaskCreate proof that a task's priority is less than
the configured max.

With the introduction of
https://github.com/FreeRTOS/FreeRTOS-Kernel/commit/9efe10b805e3516974dd5eba2d61bfa071ff8392
an assertion is added to ensure a new task's priority is less than the
confirmed max. The CBMC proof for TaskCreate needs to include this assumption
in order to not assert and fail. Since this is now enforced in the code
we can add an assumption to the proof that a task must be created with a
priority smaller than the configured max.