Don't verify full clearing of backlog
The timer task should be free to process the stop request without a
specific requirement to work through the backlog of callbacks for the
stopped timer.  The timer task should be allowed to cancel the entire
backlog, part of it, or none of it.  In other words, once the
application sends the stop request, it should not depend on receiving
any more callbacks, even backlogged ones.

The only requirement here is that the timer stop as requested.