Add mechanism for generating memory statistics report (#668)
To validate that the FreeRTOS/CI-CD-Github-Actions/memory_statistics/paths.json configuration is correct, this PR adds a CI check mechanism to generate the memory statistics report in JSON format. Additionally, the CI check also uploads the report as an artifact (for main and release-candidate branches ONLY) for user accessibility of the generated report.