VSCode launch configurations (#820)
* CORTEX_MPS2_QEMU_IAR_GCC now has VSCode tasks and launch configurations to build, run, and debug in one button. This should work on all platforms (Linux, MacOS, Windows).

* Posix_GCC Demo now has VSCode task and launch configs to build, run, debug in one button for Linux/MacOS ("lldb Launch").

It also has configuration for Windows through Ubuntu ("gdb launch"), and this will also work for Linux if the user wants to use gdb instead of lldb.

* Integrated terminal settings for Ubuntu and MSYS2 ways to run Posix Demo on Windows.

This allows Posix port demo to be built on either WSL/Ubuntu or MSYS2 on Windows.

These are absolute paths so if the user has installed Ubuntu or MSYS2 elsewhere they will need to change them.

* Improved pattern matching in Run QEMU task. Replaced usage of deprecated ${workspaceRoot} with ${workspaceFolder}.

* Split MSYS2 and Ubuntu WSL configurations

MSYS2 works better with external console, and Ubuntu WSL works with internal console. This is reflected by having two different configurations.

* Delete RTOSDemo.map

Cleanup. (This is file is built but never deleted when make-ing)

* Delete null.d

* Cleanup extranneous vscode workspace

* Documentation for VSCode launch configs

Documentation for how to run this demo on VSCode using launch configs.

* Added documentation for VSCode launch configs

Added documentation for running demo through VSCode using launch configurations.

* Removed unneeded .log files and c_cpp_properties.json in CORTEX_MPS2_QEMU_IAR_GCC/.vscode/

* Deleted unnecessary .log files from POSIX_GCC/.vscode

* Set build task problem matcher to "gcc", deault problem matcher from VSCode.

* Removed unneeded "sh -c -l" command from the default build task in CORTEX_MPS2_QEMU_IAR_GCC.

* "Build QEMU" task problem matcher finds the correct paths to problem files.

* Moved steps to "Prerequisites" section.

* Update Readme.md

Fixed markdown typo.

* Moved items to prerequisites for Posix_GCC demo.

Co-authored-by: Fan <gilbefan@f84d899204e1.ant.amazon.com>