[WIN32-MingW Demo] Add tick type width definition based on compiler type.(32bit/64bit) 32bit TickType_t is used if compiler is MinGW32. 64bit TickType_t is used if compiler is MinGW64.
Reason of change: Before this change, 32bit TickType_t is always used in MinGW demo. It is inefficient  for 64bit compiler. In addition, MinGW64 reported warnings for the cast operation between TickType_t and (void *) pointer because of different width. 64bit TickType_t should be used instead of 32bit if compiler is 64bit.