[WIN32-MSVC Demo] Add tick type width definition based on the target platform.(32bit/64bit) 32bit TickType_t is used on Win32 target. 64bit TickType_t is used on x64 target.
Reason of change: Before this change, 32bit TickType_t is always used in MSVC demo. It is inefficient  for 64bit target. In addition, compiler reported warnings for the cast operation between TickType_t and (void *) pointer if the target is x64 because of different width. 64bit TickType_t should be used instead of 32bit if target is 64bit.